Finsh rewrite on registers. Print register pair values & add padding.
This finishes the rewrite that removes the register pair variable usage.
As mentioned, the registers that the pairs "can" make up change, so we need to
mimick that. Having separated registers for this doesn't make any sense.

This commit also adds padding to the hex values printed for better clarity and
prints the BC, DE & HL pair values.