Interconnect for CPU. Fix parity on DCR. RET hack
I suspect the return address we fetch is either improperly stored in the
SP, or that since the two return addresses we create are out of scope by
the end of the CALL instruction. RET tries to grab the return address
which simply isn't in the stack pointer anymore. It should be in the
stack however..