Semi fix issue with CALL instruction
This isn't complete, currently we're only fetching the upper two bits
from memory for the subroutine return, but we also need the lower block,
i.e 18DC (we're just getting 18).
This is hacky and currently just for testing, however I realize the
problem earlier where I just wasn't modifying memory at all, so the
subroutine was lost as soon as the CALL function was out of scope.