Playing around with displaying memory values to display
This definitely isn't right. However, one thing is clear, we do have memory
values in the display module now, however, _how_ to display the bitmap is
beyond me at the moment. I'm not sure what the memory values should look like
either, which could be a good start.
Perhaps reference the memory output to another emulator for the VRAM segments.

Alternatively, we need to check the current address & match so that we're sure
we are drawing when in VRAM memory space...