Try to fix display rendering, X & Y iteration values are now changed.
This is a difficult one for me to grasp properly, and I might end up switching
back to SDL2 to be able to draw X & Y rectangles for pixels, however, at least
now the raster value gets set to 0x00 or 0xFF depending on value.

Whether or not it's correct is yet unknown.