Read bitmap into bitmap vec TODO parsing
From my brief reading on Wikipedia on the BMP format, the header file lenght
can vary, however, we could omit it entirely, or just check whether or not
the two first bytes are valid. I'm not sure what is best here.

The upside to this little side project is that it will help with rendering
down the line, as Invaders graphics too is bitmap.