Continue work on debugger renderer, add actual bitmap array
Here we've used a utility to generate an array from out bmp file, instead of
reading & parsing the bitmap file, we can troubleshoot by accessing the
array directly. What the best / cleanest implementation is I don't know.
This however presents a good way to troubleshoot forward.

Issue is minifb expects a u32 array so we need to use the byteorder crate to
create a u32 from a u8.