Use Targa format over BMP, rendering is somewhat accurate
The targa format seems superior to bmp for this purpose.
The characters rendered are actually correct although the alpha values are off.
It should be monochrome..

One issue is that there's some wrapping happening when rendering the bitmap.
It should start with W@%mqQ, but instead is rendered a 4W@...?
4 is the last glyph of the first row.

Lastly, the rendered font size is very small, that needs to be fixed.