Implement CZ & CM instructions. TODO: Better CALL instruction?
This implements the CM & CZ instructions, but it seems very redundant to do it
this way. Another way would be to improve upon the current CALL instruction &
check what the bits are & change the flags accordingly.

The issue with the current function is that it assumes your coming from an
address or calling to an address, not very flexible. :/