Debugger: Continue work on sprite sheet rendering
For some reason it renders the sheet somewhat ok if x is incremented by 3
instead of 1. It should be 1.. From what I can tell it seems to be timing issue,
and incrementing by 3 just helps it draw more? per line? Welp..

Some minor cleanup as well. At least this is somewhat progress.