Debugger: Fix sprite sheet rendering issue & simplify code.
This fixes the inverted & flipped sprite sheet rendering issue.
The reason that it was wrong was a number of things, one being that the targa
image format's 0,0 is lower left, minifb expects upper left.

Another issue was that I used file.read_to_end(), this read all of it including
the footer which I didn't know was there; it's optional..

Reading in the exact number of bytes we need & skipping the header data now
works as intended.

This also removes old bmp & tga files that are no longer used.