Debugger: Implement character rendering. Draw individual sprites
This adds support for rendering of one character from a sprite sheet.
It's not ideal as the data is flipped vs what minifb expects.
What we need to do here is write a function that essentially can takes string
in and runs it through a match statement with lookup values for each character.

This way we can print out values more easily for debugging purposes moving forward.