Display: Create 32-bit bitmap for rendering of VRAM. Set PC to 0x100 for tests
Minifb expects a 32-bit array, however, the data we have is 8-bit, so we use the
byteorder crate to convert the data into a `u32`.

Also we've set the program counter to `0x100` as some tests start here. This
needs to be removed later.

Lastly, this also adds a secondary rendering function (not yet completed)
in hopes to render out x & y values instead of the entire bitmap.