Use SP as register pair & delete duplicate code.
Also MVI now should be correct. One important part is that on LD B,$00
The second byte of immediate data is 06, this should not be read in.

For DCR we need to wrap because we decrement B before adding any data
to it ironically. At first I thought this was wrong as MVI(B) should put
data into the B register, however it should put the first immediate data
byte not the second.