Rewrite Mov to handle M register & add tests
This adds tests for HL memory access and add tests against the binary that is
essentially the same instructions as the 8080PRE HL tests.

One note about the Mov instruction is for the Mov(M,A) instruction.
M is the psuedo register and points points to HL in this case.
This causes the Cpi instruction to get the correct
immediate value to test against.