Implement Index & IndexMut for Memory
This simplifies our access to `self.memory` signifcantly as we can index the
Memory struct itself, not the vector member. By implementing `Index<u16>` we
can omit casting all indices of memory to usize, which improves readibility.

The same goes for `IndexMut<u16>` which returns a mutable reference of memory
for any operation that requires mutation within an index of memory.