Code cleanup and improve input handling
Timing and input handling is slightly improved, however not perfect.
We now use the interconnect to handle most of our data exchange between
CPU & Display etc.

Removed specified lifetimes of `<'a>` for the CPU as we no longer need
to be specific about this due to the Cpu using Memory directly instead
of through the interconnect.